Migrates to Nuxt 4 and updates ESLint config
Updates the project to Nuxt 4.

Renames files to align with the new `app/` directory structure, and configures the project to use the new Nuxt 4 conventions.

Updates the ESLint configuration to use the flat config format and includes recommended configurations.

Adds a launch configuration for debugging in VS Code.

Adds bash commands to the claude settings.